Address command palette review feedback
